requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# LLM Providers
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1
litellm==1.11.1

# Agent Tools
httpx==0.25.2
beautifulsoup4==4.12.2
duckduckgo-search==4.1.0
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6

# Utilities
aiofiles==23.2.1
